Added inputs based on indicator gradient change over past X values.
Im not sure if this works properly as I havent checked it. The gradient values seem sensible (small numbers) as the change in Y is very small comapred to X.

Perhaps this is something that should be looked into as its generating small gradient changes, could always scale the gradient values by some constant?